REF: 関数移動
